Misc changes everywhere

+++ home/config/awesome/bin/portable_poweroff.sh
@@ -1,19 +1,5 @@
 #!/usr/env/env dash
 
-sel="$(printf %s "Yes|No" | rofi -dmenu -sep "|" -p "Shutdown your system?")"
+sel="$(printf %s "Yes|No" | rofi -dmenu -sep "|" -p "Shutdown your computer?")"
 
-if [ "$sel" = "Yes" ]; then
-    OP_SYS="$(uname)"
-
-    if [ "$OP_SYS" = "Linux" ]; then
-        # For system using systemd
-        systemctl poweroff
-    elif [ "$OP_SYS" = "FreeBSD" ]; then
-        # Requires user to be in the operator group
-        shutdown -p now
-    elif [ "$OP_SYS" = "OpenBSD" ]; then
-        # Need to figure out how to shutdown without a password.
-        doas shutdown -P now
-    fi
-fi
-
+if [ "$sel" = "Yes" ]; then systemctl poweroff; fi
